refactor(ui): extract ModelRefreshMixin and finalize main.py cleanup
Final Phase 4 step:

* Move the 119-line update_from_model out of L5RMain into a dedicated
  ModelRefreshMixin in l5r/ui/refresh.py. It touches widgets across
  almost every tab (each tab mixin owns those widgets); this mixin is
  the orchestrator that pushes the current PC + rules state back into
  the UI after any state-changing action.

* Drop dead code from main.py:
  - reset_layout_geometry (never called)
  - commented-out #import sip block from the PyQt5/sip era
  - commented-out #def dump_slots debug helper
  - import l5r.dialogs / l5r.widgets aliases that are no longer used
    in main.py (their consumers are the tab mixins themselves)
  - import os (no longer used after sys.path hack removal)
  - over-specific l5r.api.* submodule imports superseded by import
    l5r.api as api

* What's left in main.py: L5RMain.__init__ (window construction +
  sinks + tab builder dispatch + signal wiring), build_ui (banner /
  tabs / settings), reset_geometry, _build_generic_page (shared by
  several tab mixins), init, connect_signals, closeEvent, and the
  module-level main() entry point.

main.py: 533 -> 374 lines.
Total Phase 4: 2551 -> 374 lines (~85% reduction).
